backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
Pillow>=10.4.0
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1